tests: Move options parser to libtests
