+++ tests/libtests/opts.c
@@ -23,20 +23,17 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
-#include "module.h"
 #include "ptr.h"
 #include "taputils.h"
 
-enum boundary_type       g_btype = CYCLE_BOUND;
-enum loop_mode           g_loop = INNER_LOOP;
-unsigned int             g_off = 0;
-const struct module*     g_module = NULL;
-unsigned int             g_num = 0;
-unsigned int             g_cycles = 10;
-unsigned int             g_nthreads = 1;
-unsigned int             g_normalize = 0;
-enum picotm_libc_cc_mode g_cc_mode = PICOTM_LIBC_CC_MODE_2PL;
-size_t                   g_txcycles = 1;
+enum boundary_type  g_btype = CYCLE_BOUND;
+enum loop_mode      g_loop = INNER_LOOP;
+unsigned int        g_off = 0;
+unsigned int        g_num = 0;
+unsigned int        g_cycles = 10;
+unsigned int        g_nthreads = 1;
+unsigned int        g_normalize = 0;
+size_t              g_txcycles = 1;
 
 static enum parse_opts_result
 opt_btype(const char* optarg)
@@ -65,25 +62,6 @@ opt_cycles(const char *optarg)
     }
 
     return PARSE_OPTS_OK;
-}
-
-static enum parse_opts_result
-opt_module(const char* optarg)
-{
-    const struct module* beg = module_list;
-    const struct module* end = module_list + number_of_modules();
-
-    while (beg < end) {
-        if (!strcmp(beg->name, optarg)) {
-            g_module = beg;
-            return PARSE_OPTS_OK;
-        }
-        ++beg;
-    }
-
-    fprintf(stderr, "Unknown module %s\n", optarg);
-
-    return PARSE_OPTS_ERROR;
 }
 
 static enum parse_opts_result
@@ -140,24 +118,6 @@ opt_off(const char* optarg)
 }
 
 static enum parse_opts_result
-opt_regular_ccmode(const char* optarg)
-{
-    static const char * const optstr[] = { "noundo", "2pl"};
-    size_t i;
-
-    for (i = 0; i < sizeof(optstr)/sizeof(optstr[0]); ++i) {
-        if (!strcmp(optstr[i], optarg)) {
-            g_cc_mode = i;
-            return PARSE_OPTS_OK;
-        }
-    }
-
-    fprintf(stderr, "unknown CC mode %s\n", optarg);
-
-    return PARSE_OPTS_ERROR;
-}
-
-static enum parse_opts_result
 opt_loop(const char* optarg)
 {
     if (!strcmp("inner", optarg)) {
@@ -201,65 +161,21 @@ opt_verbose(const char* optarg)
     return PARSE_OPTS_OK;
 }
 
-static enum parse_opts_result
-opt_help(const char* optarg)
-{
-    printf("Usage: picotm-test [options]\n"
-           "Options:\n"
-           "  -V                            About this program\n"
-           "  -h                            This help\n"
-           "  -o <number>                   Number of first test, zero upwards\n"
-           "  -m <module>                   Module name\n"
-           "  -n <number>                   Number of tests, one upwards\n"
-           "  -t <number>                   Number of concurrent threads\n"
-           "  -I <number>                   Number of iterations in transaction\n"
-           "  -R {noundo|ts|2pl|2pl-ext}    Set CC mode for file I/O\n"
-           "                                  noundo: irrevocability\n"
-           "                                  ts: timestamps\n"
-           "                                  2pl: two-phase locking\n"
-           "                                  2pl-ext: (inofficial) commit protocol for sockets\n"
-           "  -L {inner|outer}              Loop mode\n"
-           "  -b {time|cycles}              Bound for cycles\n"
-           "                                  time: bound is maximum run time in milliseconds\n"
-           "                                  cycles: cycles is maximum number of transaction runs\n"
-           "  -c <number>                   Number of cycles, aka when to stop the test\n"
-           "  -v                            Output benchmark results: '<threads> <commits> <aborts>'\n"
-           "  -N                            Normalize >commits> to transactions/second\n"
-           );
-
-    return PARSE_OPTS_EXIT;
-}
-
-static enum parse_opts_result
-opt_version(const char* optarg)
-{
-    printf("picotm test application\n");
-    printf("This software is licensed under the MIT License.\n");
-
-    return PARSE_OPTS_EXIT;
-}
+enum parse_opts_result (*parse_opt[256])(const char*) = {
+    ['I'] = opt_tx_cycles,
+    ['L'] = opt_loop,
+    ['N'] = opt_normalize,
+    ['b'] = opt_btype,
+    ['c'] = opt_cycles,
+    ['n'] = opt_num,
+    ['o'] = opt_off,
+    ['t'] = opt_nthreads,
+    ['v'] = opt_verbose
+};
 
 enum parse_opts_result
-parse_opts(int argc, char* argv[])
+parse_opts(int argc, char* argv[], const char* optstring)
 {
-    /* Parse options */
-
-    static enum parse_opts_result (* const opt[])(const char*) = {
-        ['I'] = opt_tx_cycles,
-        ['L'] = opt_loop,
-        ['N'] = opt_normalize,
-        ['R'] = opt_regular_ccmode,
-        ['V'] = opt_version,
-        ['b'] = opt_btype,
-        ['c'] = opt_cycles,
-        ['h'] = opt_help,
-        ['m'] = opt_module,
-        ['n'] = opt_num,
-        ['o'] = opt_off,
-        ['t'] = opt_nthreads,
-        ['v'] = opt_verbose
-    };
-
     if (argc < 2) {
         printf("enter `picotm-test -h` for a list of command-line options\n");
         return PARSE_OPTS_ERROR;
@@ -271,10 +187,10 @@ parse_opts(int argc, char* argv[])
         if ((c ==  '?') || (c == ':')) {
             return PARSE_OPTS_ERROR;
         }
-        if (c >= (ssize_t)arraylen(opt) || !opt[c]) {
+        if (c >= (ssize_t)arraylen(parse_opt) || !parse_opt[c]) {
             return PARSE_OPTS_ERROR;
         }
-        enum parse_opts_result res = opt[c](optarg);
+        enum parse_opts_result res = parse_opt[c](optarg);
         if (res) {
             return res;
         }
