+++ tests/modules/fildes_test.c
@@ -45,6 +45,8 @@
 #include "tempfile.h"
 #include "testhlp.h"
 
+enum picotm_libc_cc_mode g_cc_mode = PICOTM_LIBC_CC_MODE_2PL;
+
 static size_t
 MiB_to_bytes(unsigned short MiB)
 {
